Show/hide child sections when NestedSection's show/hide api is called
Previously when the NestedSection is being shown or hidden, its child section's visibilities are untouched. But this might create issues when selection or expansion is toggled ie., incorrect selection / expansion flags are set to the child section even though the parent's is not visible in the adapter. So we set visibility to all child section which will produce the correct behavior.

Note : This change was cloned from 2.x branch